+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix an issue where the aggregation wouldn't work properly with the new charts
 - Fix an issue where the user wouldn't be able to change the type of filters
 - Fix an issue where the date filters wouldn't work with "featureservice" datasets
+- Fix an issue where certain non-latin characters wouldn't be displayed correctly
 
 ## v1.4.0 - 25/02/2019
 - Prevent the editor from serializing too much information with the filters